register api

+++ src/utills/cloudinary.js
@@ -17,15 +17,17 @@ const uploadOnCloudinary = async (localFilePath,foldername) => {
     //upload the file on cloudinary
     const response = await cloudinary.uploader.upload(localFilePath, {
       resource_type: "auto",
-      folder: {foldername}, // Store images in a folder
+      folder: foldername, // Store images in a folder
     });
     // file has been uploaded successfull
     //console.log("file is uploaded on cloudinary ", response.url);
     fs.unlinkSync(localFilePath);
     return response;
   } catch (error) {
-    fs.unlinkSync(localFilePath); // remove the locally saved temporary file as the upload operation got failed
-    return null;
+    if (fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath);
+    }
+        return null;
   }
 };
